Added E2E (Cypress) for testing all parameters at once
Within this commit, I modified todo-list.component.html and todo-list.po.ts to allow for the exchange of information regarding the server-side variables of todo (status and category). Furthermore, I used these new additions to make a test that checks through and combines each parameter to ensure that parameters can work when combined.