Added the ability to add Todos and server-side testing
Within this commit, there now exists the ability for users to add todos by clicking on a + icon in the bottom-right of the /todos/ tab. There, one may enter information to create a todo.

Also, there now exists tests in TodoControllerSpec.java that looks to ensure that the various regexes and filters work as expected.